If the AllConfig class is not loaded or the deployment settings do not exist, we skip the process.

+++ src/console/architecture/createGlobalConstants/createGlobalConstantsTask.php
@@ -46,10 +46,13 @@ else
   echo 'OTRA global constants generated.', CLI_GREEN, ' ✔', END_COLOR, PHP_EOL;
 
 // On the online side
-$prodContent = str_replace($basePath, \config\AllConfig::$deployment['folder'], $content, $count);
+if (class_exists(AllConfig::class) && isset(\config\AllConfig::$deployment['folder']))
+{
+  $prodContent = str_replace($basePath, \config\AllConfig::$deployment['folder'], $content, $count);
 
-if (file_put_contents($basePath . 'config/prodConstants.php', $prodContent) === false)
-  echo CLI_RED . 'There was a problem while writing the OTRA global constants for the online side.', END_COLOR, PHP_EOL;
-else
-  echo 'OTRA global constants for the online side generated.', CLI_GREEN, ' ✔', END_COLOR, PHP_EOL;
+  if (file_put_contents($basePath . 'config/prodConstants.php', $prodContent) === false)
+    echo CLI_RED . 'There was a problem while writing the OTRA global constants for the online side.', END_COLOR, PHP_EOL;
+  else
+    echo 'OTRA global constants for the online side generated.', CLI_GREEN, ' ✔', END_COLOR, PHP_EOL;
+}
